pyproject.toml
[project]
name = "tell"
version = "0.1.0"
description = "TELL: Test-time Experiential Lifelong Learning"
readme = "readme.md"
requires-python = ">=3.12"
dependencies = [
    "arc-agi>=0.9.1",
    "python-dotenv>=1.0.0",
    "numpy>=2.3.2",
    "anthropic>=0.52.0",
    "pillow>=11.2.1",
    "pydantic>=2.11.7",
    "requests>=2.32.4",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "requests-mock>=1.12.1",
    "ruff>=0.11.4",
]

[project.optional-dependencies]
agentops = [
    "agentops>=0.4.18",
]

[tool.mypy]
strict = true
disallow_untyped_defs = true
exclude = [
    "tests",
]
plugins = []

[tool.ruff.lint]
extend-select = ["I"]
